+++ fastchat/serve/monitor/monitor.py
@@ -25,6 +25,9 @@ from fastchat.serve.monitor.elo_analysis import report_elo_analysis_results
 from fastchat.utils import build_logger, get_window_url_params_js
 
 
+from fastchat.serve.monitor.monitor_md import cat_name_to_baseline, key_to_category_name, arena_hard_title, make_default_md_1, make_default_md_2, make_arena_leaderboard_md, make_category_arena_leaderboard_md, make_full_leaderboard_md, make_leaderboard_md_live
+
+
 notebook_url = (
     "https://colab.research.google.com/drive/1KdwokPjirkTmpO_P1WByFNFiqxWQquwH"
 )
@@ -32,95 +35,21 @@ notebook_url = (
 basic_component_values = [None] * 6
 leader_component_values = [None] * 5
 
+def recompute_final_ranking(arena_df):
+    # compute ranking based on CI
+    ranking = {}
+    for i, model_a in enumerate(arena_df.index):
+        ranking[model_a] = 1
+        for j, model_b in enumerate(arena_df.index):
+            if i == j:
+                continue
+            if (
+                arena_df.loc[model_b]["rating_q025"]
+                > arena_df.loc[model_a]["rating_q975"]
+            ):
+                ranking[model_a] += 1
+    return list(ranking.values())
 
-def make_default_md_1(mirror=False):
-    link_color = "#1976D2"  # This color should be clear in both light and dark mode
-    leaderboard_md = f"""
-    # 🏆 LMSYS Chatbot Arena Leaderboard 
-    [Blog](https://lmsys.org/blog/2023-05-03-arena/) | [GitHub](https://github.com/lm-sys/FastChat) | [Paper](https://arxiv.org/abs/2403.04132) | [Dataset](https://github.com/lm-sys/FastChat/blob/main/docs/dataset_release.md) | [Twitter](https://twitter.com/lmsysorg) | [Discord](https://discord.gg/HSWAKCrnFx) | [Kaggle Competition](https://www.kaggle.com/competitions/lmsys-chatbot-arena)
-    """
-
-    return leaderboard_md
-
-
-def make_default_md_2(mirror=False):
-    mirror_str = "<span style='color: red; font-weight: bold'>This is a mirror of the live leaderboard created and maintained by the <a href='https://lmsys.org' style='color: red; text-decoration: none;'>LMSYS Organization</a>. Please link to <a href='https://leaderboard.lmsys.org' style='color: #B00020; text-decoration: none;'>leaderboard.lmsys.org</a> for citation purposes.</span>"
-    leaderboard_md = f"""
-{mirror_str if mirror else ""}
-
-LMSYS Chatbot Arena is a crowdsourced open platform for LLM evals. We've collected over 1,000,000 human pairwise comparisons to rank LLMs with the Bradley-Terry model and display the model ratings in Elo-scale.
-You can find more details in our paper. **Chatbot arena is dependent on community participation, please contribute by casting your vote!**
-
-{SURVEY_LINK}
-"""
-
-    return leaderboard_md
-
-
-def make_arena_leaderboard_md(arena_df, last_updated_time, vision=False):
-    total_votes = sum(arena_df["num_battles"]) // 2
-    total_models = len(arena_df)
-    space = "&nbsp;&nbsp;&nbsp;"
-
-    leaderboard_md = f"""
-Total #models: **{total_models}**.{space} Total #votes: **{"{:,}".format(total_votes)}**.{space} Last updated: {last_updated_time}.
-"""
-    if not vision:
-        leaderboard_md += """
-📣 **NEW!** View leaderboard for different categories (e.g., coding, long user query)! This is still in preview and subject to change.
-"""
-
-    leaderboard_md += f"""
-Code to recreate leaderboard tables and plots in this [notebook]({notebook_url}). You can contribute your vote at [chat.lmsys.org](https://chat.lmsys.org)!
-"""
-    return leaderboard_md
-
-
-def make_category_arena_leaderboard_md(arena_df, arena_subset_df, name="Overall"):
-    total_votes = sum(arena_df["num_battles"]) // 2
-    total_models = len(arena_df)
-    space = "&nbsp;&nbsp;&nbsp;"
-    total_subset_votes = sum(arena_subset_df["num_battles"]) // 2
-    total_subset_models = len(arena_subset_df)
-    leaderboard_md = f"""### {cat_name_to_explanation[name]}
-#### {space} #models: **{total_subset_models} ({round(total_subset_models/total_models *100)}%)** {space} #votes: **{"{:,}".format(total_subset_votes)} ({round(total_subset_votes/total_votes * 100)}%)**{space}
-"""
-    return leaderboard_md
-
-
-def make_full_leaderboard_md():
-    leaderboard_md = """
-Three benchmarks are displayed: **Arena Elo**, **MT-Bench** and **MMLU**.
-- [Chatbot Arena](https://chat.lmsys.org/?arena) - a crowdsourced, randomized battle platform. We use 500K+ user votes to compute model strength.
-- [MT-Bench](https://arxiv.org/abs/2306.05685): a set of challenging multi-turn questions. We use GPT-4 to grade the model responses.
-- [MMLU](https://arxiv.org/abs/2009.03300) (5-shot): a test to measure a model's multitask accuracy on 57 tasks.
-
-💻 Code: The MT-bench scores (single-answer grading on a scale of 10) are computed by [fastchat.llm_judge](https://github.com/lm-sys/FastChat/tree/main/fastchat/llm_judge).
-The MMLU scores are mostly computed by [InstructEval](https://github.com/declare-lab/instruct-eval).
-Higher values are better for all benchmarks. Empty cells mean not available.
-"""
-    return leaderboard_md
-
-
-def make_leaderboard_md_live(elo_results):
-    leaderboard_md = f"""
-# Leaderboard
-Last updated: {elo_results["last_updated_datetime"]}
-{elo_results["leaderboard_table"]}
-"""
-    return leaderboard_md
-
-
-def arena_hard_title(date):
-    arena_hard_title = f"""
-Last Updated: {date}
-
-**Arena-Hard-Auto v0.1** - an automatic evaluation tool for instruction-tuned LLMs with 500 challenging user queries curated from Chatbot Arena. 
-
-We prompt GPT-4-Turbo as judge to compare the models' responses against a baseline model (default: GPT-4-0314). If you are curious to see how well your model might perform on Chatbot Arena, we recommend trying Arena-Hard-Auto. Check out our paper for more details about how Arena-Hard-Auto works as an fully automated data pipeline converting crowdsourced data into high-quality benchmarks ->
-[[Paper](https://arxiv.org/abs/2406.11939) | [Repo](https://github.com/lm-sys/arena-hard-auto)]
-    """
-    return arena_hard_title
 
 
 def update_elo_components(
@@ -318,36 +247,6 @@ def create_ranking_str(ranking, ranking_difference):
     else:
         return f"{int(ranking)}"
 
-
-def recompute_final_ranking(arena_df):
-    # compute ranking based on CI
-    ranking = {}
-    for i, model_a in enumerate(arena_df.index):
-        ranking[model_a] = 1
-        for j, model_b in enumerate(arena_df.index):
-            if i == j:
-                continue
-            if (
-                arena_df.loc[model_b]["rating_q025"]
-                > arena_df.loc[model_a]["rating_q975"]
-            ):
-                ranking[model_a] += 1
-    return list(ranking.values())
-
-
-def highlight_top_models(df):
-    def highlight_max_rank(s):
-        # Pastel Yellow with transparency, rgba(red, green, blue, alpha)
-        highlight_color = "rgba(255, 255, 128, 0.2)"  # 50% transparent
-        if int(s["Rank* (UB)"].replace("↑", "").replace("↓", "")) == 1:
-            return [f"background-color: {highlight_color}" for _ in s]
-        else:
-            return ["" for _ in s]
-
-    # Apply and return the styled DataFrame
-    return df.apply(highlight_max_rank, axis=1)
-
-
 def get_arena_table(arena_df, model_table_df, arena_subset_df=None):
     arena_df = arena_df.sort_values(
         by=["final_ranking", "rating"], ascending=[True, False]
@@ -438,60 +337,6 @@ def get_arena_table(arena_df, model_table_df, arena_subset_df=None):
             print(f"{model_key} - {e}")
     return values
 
-
-key_to_category_name = {
-    "full": "Overall",
-    "dedup": "De-duplicate Top Redundant Queries (soon to be default)",
-    "math": "Math",
-    "if": "Instruction Following",
-    "multiturn": "Multi-Turn",
-    "coding": "Coding",
-    "hard_6": "Hard Prompts (Overall)",
-    "hard_english_6": "Hard Prompts (English)",
-    "long_user": "Longer Query",
-    "english": "English",
-    "chinese": "Chinese",
-    "french": "French",
-    "german": "German",
-    "spanish": "Spanish",
-    "russian": "Russian",
-    "japanese": "Japanese",
-    "korean": "Korean",
-    "no_tie": "Exclude Ties",
-    "no_short": "Exclude Short Query (< 5 tokens)",
-    "no_refusal": "Exclude Refusal",
-    "overall_limit_5_user_vote": "overall_limit_5_user_vote",
-    "full_old": "Overall (Deprecated)",
-}
-cat_name_to_explanation = {
-    "Overall": "Overall Questions",
-    "De-duplicate Top Redundant Queries (soon to be default)": "De-duplicate top redundant queries (top 0.1%). See details in [blog post](https://lmsys.org/blog/2024-05-17-category-hard/#note-enhancing-quality-through-de-duplication).",
-    "Math": "Math",
-    "Instruction Following": "Instruction Following",
-    "Multi-Turn": "Multi-Turn Conversation (>= 2 turns)",
-    "Coding": "Coding: whether conversation contains code snippets",
-    "Hard Prompts (Overall)": "Hard Prompts (Overall): details in [blog post](https://lmsys.org/blog/2024-05-17-category-hard/)",
-    "Hard Prompts (English)": "Hard Prompts (English), note: the delta is to English Category. details in [blog post](https://lmsys.org/blog/2024-05-17-category-hard/)",
-    "Longer Query": "Longer Query (>= 500 tokens)",
-    "English": "English Prompts",
-    "Chinese": "Chinese Prompts",
-    "French": "French Prompts",
-    "German": "German Prompts",
-    "Spanish": "Spanish Prompts",
-    "Russian": "Russian Prompts",
-    "Japanese": "Japanese Prompts",
-    "Korean": "Korean Prompts",
-    "Exclude Ties": "Exclude Ties and Bothbad",
-    "Exclude Short Query (< 5 tokens)": "Exclude Short User Query (< 5 tokens)",
-    "Exclude Refusal": 'Exclude model responses with refusal (e.g., "I cannot answer")',
-    "overall_limit_5_user_vote": "overall_limit_5_user_vote",
-    "Overall (Deprecated)": "Overall without De-duplicating Top Redundant Queries (top 0.1%). See details in [blog post](https://lmsys.org/blog/2024-05-17-category-hard/#note-enhancing-quality-through-de-duplication).",
-}
-cat_name_to_baseline = {
-    "Hard Prompts (English)": "English",
-}
-
-
 def update_leaderboard_df(arena_table_vals):
     elo_datarame = pd.DataFrame(
         arena_table_vals,
@@ -533,7 +378,7 @@ def update_leaderboard_df(arena_table_vals):
             )
             for v in s
         ]
-
+    
     return elo_datarame.style.apply(highlight_max, subset=["Rank* (UB)"]).apply(
         highlight_rank_max, subset=["Delta"]
     )
@@ -612,10 +457,10 @@ def build_arena_tab(
             arena_values = gr.Dataframe(
                 headers=[
                     "Rank* (UB)",
-                    "🤖 Model",
-                    "⭐ Arena Score",
-                    "📊 95% CI",
-                    "🗳️ Votes",
+                    "Model",
+                    "Arena Score",
+                    "95% CI",
+                    "Votes",
                     "Organization",
                     "License",
                     "Knowledge Cutoff",
@@ -691,10 +536,10 @@ def build_arena_tab(
     elo_display_df = gr.Dataframe(
         headers=[
             "Rank* (UB)",
-            "🤖 Model",
-            "⭐ Arena Elo",
-            "📊 95% CI",
-            "🗳️ Votes",
+            "Model",
+            "Arena Elo",
+            "95% CI",
+            "Votes",
             "Organization",
             "License",
             "Knowledge Cutoff",
@@ -802,6 +647,84 @@ def build_full_leaderboard_tab(elo_results, model_table_df, model_to_score):
         wrap=True,
     )
 
+def get_arena_category_table(elo_results, model_table_df, categories):
+    print("Building category leaderboard table")
+    print(elo_results.keys())
+    new_category_dfs= []
+    for category in categories:
+        category_df = elo_results[category]["leaderboard_table_df"]
+        category_df[key_to_category_name[category]] = recompute_final_ranking(category_df)
+        new_category_dfs.append(category_df[[key_to_category_name[category]]])
+    category_df = pd.concat(new_category_dfs, axis=1)
+    print(category_df.columns[0])
+    category_df = category_df.sort_values(by=category_df.columns[0], ascending=True)
+    
+    def get_model_name(model_key):
+        try:
+            model_name = model_table_df[model_table_df["key"] == model_key][
+                        "Model"
+                    ].values[0]
+            return model_name
+        except:
+            return None
+        
+    category_df["Model"] = category_df.index
+    category_df["Model"] = category_df["Model"].apply(get_model_name)
+    # remove models that are not in the model table
+    category_df = category_df[category_df["Model"].notnull()]
+    # Move Model column to the front
+    cols = category_df.columns.tolist()
+    cols = cols[-1:] + cols[:-1]
+    category_df = category_df[cols]
+
+    category_df = pd.DataFrame(
+        category_df.values,
+        columns=["Model"] + [key_to_category_name[k] for k in categories],
+    )
+
+    def highlight_top_3(s):
+        return [
+            (
+                "background-color: rgba(255, 215, 0, 0.3); text-align: center; font-size: 110%"
+                if int(v) == 1
+                else "background-color: rgba(211, 211, 211, 0.4); text-align: center; font-size: 110%"
+                if int(v) == 2
+                else "background-color: rgba(209, 139, 71, 0.2); text-align: center; font-size: 110%"
+                if int(v) == 3
+                else "text-align: center; font-size: 110%"
+            )
+            for v in s
+        ]
+    
+    return category_df.style.apply(highlight_top_3, subset=[key_to_category_name[k] for k in categories])
+
+def build_category_leaderboard_tab(elo_results, model_table_df, categories, categories_width):
+    
+#     gr.Markdown(
+#         f"""
+# **We've updated the leaderboard to show model rank (UB) across categories for better performance comparison. For more categories, statistics, and model info, see the Arena tab.**
+# """,
+#         elem_id="leaderboard_markdown",
+#     )
+#     last_updated_time = elo_results["full"]["last_updated_datetime"].split(" ")[0]
+#     gr.Markdown(make_arena_leaderboard_md(elo_results["full"]["leaderboard_table_df"], last_updated_time), elem_id="leaderboard_markdown")
+    full_table_vals = get_arena_category_table(elo_results, model_table_df, categories)
+    gr.Dataframe(
+        headers=["Model"] + [key_to_category_name[k] for k in categories],
+        datatype=["markdown"] + ["str" for k in categories],
+        value=full_table_vals,
+        elem_id="full_leaderboard_dataframe",
+        column_widths=[250] + categories_width, # IMPORTANT: THIS IS HARDCODED WITH THE CURRENT CATEGORIES
+        height=800,
+        wrap=True,
+    )
+
+
+selected_categories = ['full', 'coding', 'if', 'math', 'hard_6', 'multiturn', 'long_user', 'no_refusal']
+selected_categories_width = [85, 75, 120, 75, 125, 110, 100, 100]
+
+language_categories = ['english', 'chinese', 'german', 'french', 'spanish', 'russian', 'japanese', 'korean']
+language_categories_width = [100] * len(language_categories)
 
 def build_leaderboard_tab(
     elo_results_file,
@@ -837,14 +760,48 @@ def build_leaderboard_tab(
         model_table_df = pd.DataFrame(data)
 
         with gr.Tabs() as tabs:
-            with gr.Tab("Arena", id=0):
+            with gr.Tab("Ranking Breakdown", id=0):
+                gr.Markdown(
+                    f"""
+                    <div style="text-align: center; font-weight: bold;">
+                        We've updated the leaderboard to show model rank (UB) across categories to provide a more holistic comparison. Check out the Arena tab for more categories, statistics, and model info.
+                    </div>
+                    """,
+                )
+                last_updated_time = elo_results_text["full"]["last_updated_datetime"].split(" ")[0]
+                gr.Markdown(make_arena_leaderboard_md(elo_results_text["full"]["leaderboard_table_df"], last_updated_time), elem_id="leaderboard_markdown")
+                gr.Markdown("""<span style='text-decoration:underline; font-size: 125%;'>Task Leaderboard</span>""")
+                gr_plots = build_category_leaderboard_tab(
+                    elo_results_text,
+                    model_table_df,
+                    selected_categories,
+                    selected_categories_width,
+                )
+                gr.Markdown("""<span style='text-decoration:underline; font-size: 125%;'>Language Leaderboard</span>""")
+                build_category_leaderboard_tab(
+                    elo_results_text,
+                    model_table_df,
+                    language_categories,
+                    language_categories_width,
+                )
+                gr.Markdown(
+                    f"""
+            ***Rank (UB)**: model's ranking (upper-bound), defined by one + the number of models that are statistically better than the target model.
+            Model A is statistically better than model B when A's lower-bound score is greater than B's upper-bound score (in 95% confidence interval).
+            See Figure 1 below for visualization of the confidence intervals of model scores.
+
+            Note: in each category, we exclude models with fewer than 300 votes as their confidence intervals can be large.
+            """,
+                    elem_id="leaderboard_markdown",
+                )
+            with gr.Tab("Arena", id=1):
                 gr_plots = build_arena_tab(
                     elo_results_text,
                     model_table_df,
                     default_md,
                     show_plot=show_plot,
                 )
-            with gr.Tab("📣 NEW: Arena (Vision)", id=1):
+            with gr.Tab("Arena (Vision)", id=2):
                 build_arena_tab(
                     elo_results_vision,
                     model_table_df,
@@ -853,7 +810,7 @@ def build_leaderboard_tab(
                     show_plot=show_plot,
                 )
             if arena_hard_leaderboard is not None:
-                with gr.Tab("Arena-Hard-Auto", id=2):
+                with gr.Tab("Arena-Hard-Auto", id=3):
                     dataFrame = arena_hard_process(
                         leaderboard_table_file, arena_hard_leaderboard
                     )
@@ -890,7 +847,7 @@ def build_leaderboard_tab(
                         column_widths=[70, 190, 80, 80, 90, 150],
                     )
 
-            with gr.Tab("Full Leaderboard", id=3):
+            with gr.Tab("Full Leaderboard", id=4):
                 build_full_leaderboard_tab(
                     elo_results_text, model_table_df, model_to_score
                 )
@@ -929,13 +886,13 @@ def build_leaderboard_tab(
 
     return [md_1] + gr_plots
 
-
 def build_demo(elo_results_file, leaderboard_table_file, arena_hard_leaderboard):
     from fastchat.serve.gradio_web_server import block_css
 
     text_size = gr.themes.sizes.text_lg
     # load theme from theme.json
-    theme = gr.themes.Default.load("theme.json")
+    # theme = gr.themes.Default.load("theme.json")
+    theme=gr.themes.Default()
     # set text size to large
     theme.text_size = text_size
     theme.set(
